Changes some stuff
