+++ .env.example
@@ -4,3 +4,12 @@ INFLUX_ORGANIZATION=
 INFLUX_BUCKET=
 INFLUX_HOST=
 
+# Database configs
+MONGO_USERNAME=
+MONGO_PASSWORD=
+MONGO_DB=
+MONGO_HOST=
+
+SERVER_PORT=
+
+JWT_SIGNING_SECRET=
