+++ middlewares/jwtAuth.js
@@ -1,4 +1,4 @@
-const accessTokenSecret = 'thisshouldbestoredintheenvbutisokayfornow';
+const jwt = require('jsonwebtoken');
 
 const authenticateJWT = (req, res, next) => {
     const authHeader = req.headers.authorization;
@@ -6,7 +6,7 @@ const authenticateJWT = (req, res, next) => {
     if (authHeader) {
         const token = authHeader.split(' ')[1];
 
-        jwt.verify(token, accessTokenSecret, (err, user) => {
+        jwt.verify(token, process.env.JWT_SIGNING_SECRET, (err, user) => {
             if (err) {
                 return res.sendStatus(403);
             }
